requirements.txt
absl-py==0.9.0
arrow==0.15.5
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
basemap @ file:///C:/Users/LSG/Documents/Professionnalisation/StageFAA/Hands-on_ML/env/basemap-1.2.1-cp37-cp37m-win_amd64.whl
bleach==3.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.3
cycler==0.10.0
Cython==0.29.16
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
future==0.18.2
gast==0.2.2
geos==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gym==0.17.1
h5py==2.10.0
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.3
ipython==7.11.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.2.0
proj==0.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyglet==1.5.0
Pygments==2.5.2
pyparsing==2.4.6
pyproj==2.6.1.post1
pyrsistent==0.15.6
pyshp==2.1.0
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
